WIP - separate minimal settings from wrapper settings
