+++ src/main/java/de/embl/cba/elastixwrapper/settings/TransformixSettings.java
@@ -1,0 +1,6 @@
+package de.embl.cba.elastixwrapper.settings;
+
+public class TransformixSettings extends Settings {
+    public String transformationFilePath;
+    public String movingImageFilePath;
+}
