+++ src/main/java/de/embl/cba/elastixwrapper/settings/TransformixWrapperSettings.java
@@ -1,0 +1,18 @@
+package de.embl.cba.elastixwrapper.settings;
+
+import java.io.File;
+
+public class TransformixWrapperSettings extends TransformixSettings {
+
+    public enum OutputModality
+    {
+        Show_images,
+        Save_as_tiff,
+        Save_as_BigDataViewer_xml_h5
+    }
+
+    public OutputModality outputModality;
+    public File outputFile;
+    public File outputDirectory;
+    public String transformationOutputFilePath;
+}
