+++ src/main/java/de/embl/cba/elastixwrapper/utils/CommandLineUtils.java
@@ -1,0 +1,93 @@
+package de.embl.cba.elastixwrapper.utils;
+
+import ij.IJ;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.Map;
+
+import static de.embl.cba.elastixwrapper.utils.Utils.saveStringToFile;
+import static org.scijava.util.PlatformUtils.*;
+
+public class CommandLineUtils {
+    public static String createExecutableShellScript( String elastixOrTransformix )
+    {
+        if ( isMac() || isLinux() )
+        {
+            String executablePath = tmpDir
+                    + File.separator + "run_" + elastixOrTransformix + ".sh";
+
+            String binaryPath = elastixDirectory + File.separator + "bin" + File.separator + elastixOrTransformix;
+
+            if( ! new File( binaryPath ).exists() )
+                Utils.logErrorAndExit( settings, "Elastix file does not exist: " + binaryPath );
+
+            String shellScriptText = getScriptText( elastixOrTransformix );
+
+            saveStringToFile( shellScriptText, executablePath );
+
+            makeExecutable( executablePath );
+
+            return executablePath;
+
+        }
+        else if ( isWindows() )
+        {
+            setElastixSystemPathForWindowsOS();
+
+            String binaryPath = elastixDirectory + File.separator + elastixOrTransformix + ".exe";
+
+            if ( ! new File( binaryPath ).exists() )
+                Utils.logErrorAndExit( settings, "Elastix file does not exist: " + binaryPath );
+
+            return binaryPath;
+        }
+        else
+        {
+            Utils.logErrorAndExit( settings, "Could not detect operating system!" );
+            return null;
+        }
+
+    }
+
+    private String getScriptText( String elastixOrTransformix )
+    {
+        String shellScriptText = "";
+        shellScriptText += "#!/bin/bash\n";
+        shellScriptText += "ELASTIX_PATH=" + elastixDirectory + "\n";
+
+        if ( isMac() )
+        {
+            shellScriptText += "export DYLD_LIBRARY_PATH=$ELASTIX_PATH/lib/\n";
+        }
+        else if ( isLinux() )
+        {
+            shellScriptText += "export LD_LIBRARY_PATH=$ELASTIX_PATH/lib/\n";
+        }
+
+        shellScriptText += "$ELASTIX_PATH/bin/" + elastixOrTransformix +" $@\n";
+        return shellScriptText;
+    }
+
+    private void setElastixSystemPathForWindowsOS()
+    {
+        ProcessBuilder pb = new ProcessBuilder();
+        Map<String, String> env = pb.environment();
+        env.put( "PATH", elastixDirectory + ":$PATH");
+    }
+
+    private void makeExecutable( String executablePath )
+    {
+        try
+        {
+            Utils.waitOneSecond();
+            Runtime.getRuntime().exec("chmod +x " + executablePath );
+            Utils.waitOneSecond();
+        }
+        catch ( IOException e )
+        {
+            IJ.log( "Could not make file executable: " + executablePath );
+            e.printStackTrace();
+        }
+    }
+}
