+++ src/test/java/ExampleElastixAPI.java
@@ -1,4 +1,4 @@
-import de.embl.cba.elastixwrapper.elastix.ElastixSettings;
+import de.embl.cba.elastixwrapper.settings.ElastixWrapperSettings;
 import de.embl.cba.elastixwrapper.elastix.ElastixWrapper;
 
 import net.imagej.ImageJ;
@@ -10,12 +10,12 @@ public class ExampleElastixAPI
 		final ImageJ ij = new ImageJ();
 		ij.ui().showUI();
 
-		ElastixSettings settings = new ElastixSettings();
+		ElastixWrapperSettings settings = new ElastixWrapperSettings();
 
 		settings.logService = ij.log();
 		settings.elastixDirectory = "/Applications/elastix_macosx64_v4.8" ;
 		settings.tmpDir = "/Users/tischer/Desktop/elastix-tmp";
-		settings.transformationType = ElastixSettings.EULER;
+		settings.transformationType = ElastixWrapperSettings.EULER;
 		settings.fixedImageFilePath = getImageFilePath( "test-data/fluo01/ellipsoid-horizontal-dxyz200nm.tif" );
 		settings.movingImageFilePath = getImageFilePath( "test-data/fluo01/ellipsoid-at45degrees-dxyz200nm.tif" );
 		settings.downSamplingFactors = "10 10 10";
