+++ src/test/java/tests/TestTransformix.java
@@ -1,6 +1,6 @@
 package tests;
 
-import de.embl.cba.elastixwrapper.elastix.ElastixSettings;
+import de.embl.cba.elastixwrapper.settings.ElastixWrapperSettings;
 import de.embl.cba.elastixwrapper.elastix.ElastixWrapper;
 import net.imagej.ImageJ;
 import org.junit.Test;
@@ -12,7 +12,7 @@ public class TestTransformix
 	@Test
 	public void registerSingleChannelImageAndSaveAsTiff()
 	{
-		ElastixSettings settings = new ElastixSettings();
+		ElastixWrapperSettings settings = new ElastixWrapperSettings();
 
 		final ImageJ ij = new ImageJ();
 
@@ -22,7 +22,7 @@ public class TestTransformix
 		settings.movingImageFilePath = "/Users/tischer/Documents/fiji-plugin-elastixWrapper/src/test/resources/test-data/fluo01/ellipsoid-at45degrees-dxyz200nm.tif";
 		settings.transformationFilePath = "/Users/tischer/Documents/fiji-plugin-elastixWrapper/src/test/resources/test-data/fluo01/tmp/TransformParameters.0.txt";
 
-		settings.outputModality = ElastixSettings.OUTPUT_MODALITY_SAVE_AS_TIFF;
+		settings.outputModality = ElastixWrapperSettings.OUTPUT_MODALITY_SAVE_AS_TIFF;
 		settings.outputFile = new File( "/Users/tischer/Documents/fiji-plugin-elastixWrapper/src/test/resources/test-data/fluo01/transformed-ellipsoid" );
 
 		final ElastixWrapper elastixWrapper = new ElastixWrapper( settings );
