+++ src/test/java/users/rachel/RachelParapodRegistration.java
@@ -1,7 +1,7 @@
 package users.rachel;
 
 import bdv.util.Bdv;
-import de.embl.cba.elastixwrapper.elastix.ElastixSettings;
+import de.embl.cba.elastixwrapper.settings.ElastixWrapperSettings;
 import de.embl.cba.elastixwrapper.elastix.ElastixWrapper;
 import net.imagej.ImageJ;
 
@@ -12,7 +12,7 @@ public class RachelParapodRegistration
 		final ImageJ ij = new ImageJ();
 		ij.ui().showUI();
 
-		ElastixSettings settings = new ElastixSettings();
+		ElastixWrapperSettings settings = new ElastixWrapperSettings();
 
 		settings.logService = ij.log();
 		settings.elastixDirectory = "/Applications/elastix_macosx64_v4.8" ;
@@ -26,7 +26,7 @@ public class RachelParapodRegistration
 
 		settings.initialTransformationFilePath = "/Users/tischer/Documents/rachel-mellwig-em-prospr-registration/data/ganglion-segmentation/amira-transform.txt";
 
-		settings.transformationType = ElastixSettings.AFFINE;
+		settings.transformationType = ElastixWrapperSettings.AFFINE;
 		settings.downSamplingFactors = "10 10 10; 2, 2, 2";
 		// settings.movingMaskPath = "";
 		// settings.bSplineGridSpacing = "50 50 50";
